wip: global IC prioritization
