+++ distage/distage-core/src/main/scala/izumi/distage/planning/PlannerDefaultImpl.scala
@@ -1,26 +1,21 @@
 package izumi.distage.planning
 
-import izumi.distage.model.definition.Axis.AxisChoice
-import izumi.distage.model.definition.BindingTag.AxisTag
-import izumi.distage.model.definition.conflicts.ConflictResolutionError.{ConflictingAxisChoices, ConflictingDefs, UnsolvedConflicts}
-import izumi.distage.model.definition.conflicts.{ConflictResolutionError, MutSel}
+import izumi.distage.model.definition.conflicts.MutSel
 import izumi.distage.model.definition.errors.DIError
-import izumi.distage.model.definition.errors.DIError.{ConflictResolutionFailed, LoopResolutionError, VerificationError}
-import izumi.distage.model.definition.{Activation, Binding, ModuleBase}
-import izumi.distage.model.exceptions.planning.{ConflictResolutionException, DIBugException}
+import izumi.distage.model.definition.errors.DIError.{ConflictResolutionFailed, LoopResolutionError}
+import izumi.distage.model.definition.{Activation, ModuleBase}
 import izumi.distage.model.exceptions.InjectorFailed
-import izumi.distage.model.plan.ExecutableOp.{ImportDependency, InstantiationOp, SemiplanOp}
+import izumi.distage.model.exceptions.planning.{ConflictResolutionException, DIBugException}
 import izumi.distage.model.plan.*
+import izumi.distage.model.plan.ExecutableOp.{ImportDependency, InstantiationOp, SemiplanOp}
 import izumi.distage.model.plan.operations.OperationOrigin
 import izumi.distage.model.plan.operations.OperationOrigin.EqualizedOperationOrigin
-import izumi.distage.model.plan.repr.KeyMinimizer
 import izumi.distage.model.planning.*
 import izumi.distage.model.reflection.DIKey
 import izumi.distage.model.{Planner, PlannerInput}
 import izumi.distage.planning.solver.{PlanSolver, SemigraphSolver}
 import izumi.fundamentals.graphs.struct.IncidenceMatrix
 import izumi.fundamentals.graphs.{DG, GraphMeta}
-import izumi.fundamentals.platform.strings.IzString.*
 
 import scala.annotation.nowarn
 
@@ -33,7 +28,7 @@ class PlannerDefaultImpl(
   resolver: PlanSolver,
 ) extends Planner {
 
-  import scala.collection.compat._
+  import scala.collection.compat.*
 
   override def planSafe(input: PlannerInput): Either[List[DIError], DIPlan] = {
     planNoRewriteSafe(input.copy(bindings = rewrite(input.bindings)))
@@ -168,47 +163,15 @@ class PlannerDefaultImpl(
     DG.fromPred(IncidenceMatrix(plan.predecessors.links ++ allImports), fullMeta)
   }
 
-  def formatError(e: VerificationError): String = e match {
-    case VerificationError.BUG_PlanIndexIsBroken(badIndex) =>
-      s"BUG: plan index keys are inconsistent with operations: $badIndex"
-    case VerificationError.BUG_PlanIndexHasUnrequiredOps(unreferencedInGraph) =>
-      s"BUG: plan index has operations not referenced in dependency graph: $unreferencedInGraph"
-    case VerificationError.BUG_PlanMatricesInconsistent(plan) =>
-      s"BUG: predcessor matrix in the plan is not equal to transposed successor matrix: $plan"
-    case VerificationError.BUG_InitWithoutProxy(missingProxies) =>
-      s"BUG: Cannot finish the plan, there are missing MakeProxy operations: $missingProxies!"
-    case VerificationError.BUG_ProxyWithoutInit(missingInits) =>
-      s"BUG: Cannot finish the plan, there are missing InitProxy operations: $missingInits!"
-    case VerificationError.PlanReferencesMissingOperations(missingInOpsIndex) =>
-      s"Plan graph references missing operations: ${missingInOpsIndex.niceList()}"
-    case VerificationError.MissingRefException(missing, _) =>
-      s"Plan is broken, there the following keys are declared as dependencies but missing from the graph: ${missing.niceList()}"
-    case VerificationError.MissingRoots(roots) =>
-      s"There are no operations for the following plan roots: ${roots.niceList()}"
-  }
-  def formatError(e: LoopResolutionError): String = e match {
-    case LoopResolutionError.BUG_NotALoopMember(badOp) =>
-      s"BUG: ${badOp.target} is not an operation which expected to be a user of a cycle"
-
-    case LoopResolutionError.BUG_UnableToFindLoop(predcessors) =>
-      s"BUG: Failed to break circular dependencies, loop detector failed on matrix $predcessors which is expected to contain a loop"
-
-    case LoopResolutionError.BUG_BestLoopResolutionIsNotSupported(op) =>
-      s"BUG: Failed to break circular dependencies, best candidate ${op.target} is not a proxyable operation: $op"
-
-    case LoopResolutionError.BestLoopResolutionCannotBeProxied(op) =>
-      s"Failed to break circular dependencies, best candidate ${op.target} is not proxyable (final?): $op"
-  }
-
   // TODO: we need to completely get rid of exceptions, this is just some transitional stuff
   protected[this] def throwOnError(activation: Activation, issues: List[DIError]): Nothing = {
     val conflicts = issues.collect { case c: ConflictResolutionFailed => c }
     if (conflicts.nonEmpty) {
       throwOnConflict(activation, conflicts)
     }
-    import izumi.fundamentals.platform.strings.IzString._
+    import izumi.fundamentals.platform.strings.IzString.*
 
-    val loops = issues.collect { case e: LoopResolutionError => formatError(e) }.niceList()
+    val loops = issues.collect { case e: LoopResolutionError => DIError.formatError(e) }.niceList()
     if (loops.nonEmpty) {
       throw new InjectorFailed(
         s"""Injector failed unexpectedly. List of issues: $loops
@@ -217,7 +180,7 @@ class PlannerDefaultImpl(
       )
     }
 
-    val inconsistencies = issues.collect { case e: DIError.VerificationError => formatError(e) }.niceList()
+    val inconsistencies = issues.collect { case e: DIError.VerificationError => DIError.formatError(e) }.niceList()
     if (inconsistencies.nonEmpty) {
       throw new InjectorFailed(
         s"""Injector failed unexpectedly. List of issues: $loops
@@ -230,7 +193,7 @@ class PlannerDefaultImpl(
   }
   protected[this] def throwOnConflict(activation: Activation, issues: List[ConflictResolutionFailed]): Nothing = {
     val rawIssues = issues.map(_.error)
-    val issueRepr = rawIssues.map(formatConflict(activation)).mkString("\n", "\n", "")
+    val issueRepr = rawIssues.map(DIError.formatConflict(activation)).mkString("\n", "\n", "")
 
     throw new ConflictResolutionException(
       s"""Found multiple instances for a key. There must be exactly one binding for each DIKey. List of issues:$issueRepr
@@ -239,73 +202,6 @@ class PlannerDefaultImpl(
        """.stripMargin,
       rawIssues,
     )
-  }
-
-  protected[this] def formatConflict(activation: Activation)(conflictResolutionError: ConflictResolutionError[DIKey, InstantiationOp]): String = {
-    conflictResolutionError match {
-      case ConflictingAxisChoices(issues) =>
-        val printedActivationSelections = issues.map {
-          case (axis, choices) => s"axis: `$axis`, selected: {${choices.map(_.value).mkString(", ")}}"
-        }
-        s"""Multiple axis choices selected for axes, only one choice must be made selected for an axis:
-           |
-           |${printedActivationSelections.niceList().shift(4)}""".stripMargin
-
-      case ConflictingDefs(defs) =>
-        defs
-          .map {
-            case (k, nodes) =>
-              conflictingAxisTagsHint(
-                key = k,
-                activeChoices = activation.activeChoices.values.toSet,
-                ops = nodes.map(_._2.meta.origin.value),
-              )
-          }.niceList()
-
-      case UnsolvedConflicts(defs) =>
-        defs
-          .map {
-            case (k, axisBinds) =>
-              s"""Conflict resolution failed for key:
-                 |
-                 |   - ${k.asString}
-                 |
-                 |   Reason: Unsolved conflicts.
-                 |
-                 |   Candidates left: ${axisBinds.niceList().shift(4)}""".stripMargin
-          }.niceList()
-    }
-  }
-
-  protected[this] def conflictingAxisTagsHint(
-    key: MutSel[DIKey],
-    activeChoices: Set[AxisChoice],
-    ops: Set[OperationOrigin],
-  ): String = {
-    val keyMinimizer = KeyMinimizer(
-      ops.flatMap(_.foldPartial[Set[DIKey]](Set.empty, { case b: Binding.ImplBinding => Set(DIKey.TypeKey(b.implementation.implType)) }))
-      + key.key,
-      colors = false,
-    )
-    val axisValuesInBindings = ops.iterator.collect { case d: OperationOrigin.Defined => d.binding.tags }.flatten.collect { case AxisTag(t) => t }.toSet
-    val alreadyActiveTags = activeChoices.intersect(axisValuesInBindings)
-    val candidates = ops.iterator
-      .map {
-        op =>
-          val bindingTags = op.fold(Set.empty[AxisChoice], _.tags.collect { case AxisTag(t) => t })
-          val conflicting = axisValuesInBindings.diff(bindingTags)
-          val implTypeStr = op.foldPartial("", { case b: Binding.ImplBinding => keyMinimizer.renderType(b.implementation.implType) })
-          s"$implTypeStr ${op.toSourceFilePosition} - required: {${bindingTags.mkString(", ")}}, conflicting: {${conflicting.mkString(", ")}}, active: {${alreadyActiveTags
-            .mkString(", ")}}"
-      }.niceList().shift(4)
-
-    s"""Conflict resolution failed for key:
-       |
-       |   - ${keyMinimizer.renderKey(key.key)}
-       |
-       |   Reason: Conflicting definitions available without a disambiguating axis choice.
-       |
-       |   Candidates left:$candidates""".stripMargin
   }
 
 }
